Avoid NPE when RedirectHandler is cleared during incognito dialog
In handleFallbackUrl, which is called when the user chooses not to leave
incognito, we check the RedirectHandler to see if the fallback URL can
leave Chrome. However, if the user backgrounds Chrome while the leave
incognito dialog is up, the RedirectHandler is cleared and this check
causes a NPE.

There shouldn't be any fingerprinting concerns in this case (which is
one of the things the crashing check was preventing) as there's both a
blocking dialog in the way, preventing passive fingerprinting, and the
redirectHandler is cleared so the clobbering navigation won't have a
gesture and so cannot redirect out of Chrome anyways.

Bug: 1302566
Change-Id: I1031e6dc55f5b9bded2e53effea648e94b2a0288
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3503493
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977660}