Fix flaky TimeFormatUtilTest.GetViewedItemText
Time should be passed using base::Time, passing it as an int lost
precision and resulted in flaky results. Storing it as an int is fine
but that should be an implementation detail at the storage site,
understanding that we loose precision there as well but being okay
with that.

Tweaked the "Viewed in the past month" test case as it was
effectively testing (now - base::Days(30)) < (now - base::Days(30)).
The only reason this passed most of the time was the loss of
precision making the recomputed value land a second short...
Flakily being == when the microseconds representation of Now() luckily
had enough 0's in its tail to not lose precision...

Fixed: 1302663
Change-Id: I0ab90fe66bc4c1927429c3fbaa9cad91da1e2db6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3498477
Reviewed-by: Mohamad Ahmadi <mahmadi@chromium.org>
Commit-Queue: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977664}