[BlockFragmentation] Increase cache hit-rate for new formatting contexts
Previously we'd miss the cache whenever we caching logic didn't have
a |bfc_block_offset|.

For blocks this is correct as we might have a self-collapsing element
where we haven't decided the final position for.

However new-FCs also don't have a specified |bfc_block_offset|. For
these cases using FragmentainerOffsetAtBfc() directly is correct.

With this patch we can now hit the cache between the initial-pass, and
layout-pass for new-FCs.

Bug: 829028
Change-Id: I4aa757367dc7d7a4943a792b26cd33f64c4825a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3425795
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977698}