Add measurement of XRequestedWith API for WebView
Refactoring the metrics measurement currently in use for dark mode to
also encapsulate the metrics for other settings, in this case the
X-Requested-With HTTP header settings.

Adding measurement of the header type that is being set at request time,
including if something is setting an explicit header.

Change-Id: I3196bf8a8f78c7d108a63d09b0fe0b26fc09d77f
Bug: 1295213
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3472493
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Caitlin Fischer <caitlinfischer@google.com>
Commit-Queue: Peter Pakkenberg <pbirk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977700}