WebView: Update instrumentation test to exercise additional file parameters.
In order to verify the full functionality of the showFileChooser API,
additional tests need to be added beyond the existing test suite which
only exercises the single and multiple file choice modes for a
file input type.

Adding tests for the accepted types and whether capture is enabled
or disabled.

Bug: 1214124
Change-Id: I1a02fac5ec880c6c31e05faa347771e6f9c00d21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3482455
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Commit-Queue: Adam Walls <avvall@google.com>
Cr-Commit-Position: refs/heads/main@{#977708}