Revert "Drop duplicate and out of order VideoFrames in MSTP"
This reverts commit 6ec854afe25f722c9e8d1f1ac4e9c7b5a4829f4a.

Reason for revert: Causes frames to be dropped for a long time after the
source of a track gets switched to a new one with timestamps which reset
to 0 - eg due to changing camera resolution via a applyConstraints call
on windows, or using the "Share this tab instead" button.

Note: Partial revert to leave the histogram config and WPT in place.

Original change's description:
> Drop duplicate and out of order VideoFrames in MSTP
>
> Drop videoframes which arrive on a track with duplicate or out of order timestamps, to make
> it easier for apps to handle frames without worrying about order.
>
> Bug: 1271175
> Change-Id: I20a7312a31d8122c77dccde6d5064b27f0637db3
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3304220
> Commit-Queue: Tony Herre <toprice@chromium.org>
> Reviewed-by: Ben Wagner aka dogben <benjaminwagner@google.com>
> Reviewed-by: Johannes Kron <kron@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#946371}

Bug: 1271175, 1301897, 1287823
Change-Id: I127a0366dca83bc4ca5f31e0c218f919919af5f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3497696
Reviewed-by: Ben Wagner <benjaminwagner@google.com>
Commit-Queue: Tony Herre <toprice@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977730}