UAF in ViewsAXTreeManager follow up
This is a follow-up to
https://chromium-review.googlesource.com/c/chromium/src/+/3346380
...in which I added some null checks that fixed the UAF locally.

The bug opener indicated that this UAF can still be hit after that change. They
have provided several excellent repro's that will trigger it.

The core of the issue here is that the 'LoadComplete' event that gets fired upon
construction of ViewsAXTreeManager cannot be fired synchronously, as the
AXAuraObjectCache has not added the wrapper from the cache at that point.

This trimmed call stack demonstrates the issue :

00 ViewAccessibility::ViewAccessibility view: 0x1cdea37b220
01 ViewAXPlatformNodeDelegate::ViewAXPlatformNodeDelegate
...
04 ViewAccessibility::Create
05 View::GetViewAccessibility
06 AXViewObjWrapper::GetUniqueId
...
08 AXAuraObjCache::GetOrCreate
09 ViewsAXTreeManager::OnViewEvent
0a AXEventManager::NotifyViewEvent
0b View::NotifyAccessibilityEvent (LoadComplete)
0c ViewsAXTreeManager::ViewsAXTreeManager
...
0e ViewAccessibility::ViewAccessibility this: view: 0x1cdea37b220
0f ViewAXPlatformNodeDelegate::ViewAXPlatformNodeDelegate
...
12 ViewAccessibility::Create
13 View::GetViewAccessibility
14 AXViewObjWrapper::GetUniqueId
15 AXAuraObjCache::CreateInternal<views::AXViewObjWrapper,views::View>
16 AXAuraObjCache::GetOrCreate
17 ViewsAXTreeManager::OnViewEvent
...
21 Widget::Init

Reading from the bottom up, you can see the expected 'ViewAccessibility::Create'
call upon creating the root Widget. However, on frame 0b,
View::NotifyAccessibilityEvent (LoadComplete) calls into
AXViewObjWrapper::GetUniqueId, which creates *another* ViewAccessibility, for
the same View as above (in this case, 0x1cdea37b220). There should never be
two ViewAccessibility nodes for the same view, and why the UAF is possible.

My fix here is to simply fire LoadComplete asynchronously. That allows the stack
to unwind such that the first call to |AXAuraObjCache::GetOrCreate| adds the
wrapper to the cache, so when LoadComplete is finally called, it is able to
retrieve it instead of creating a second one.

A unit test and DCHECK were added to enforce this. The unit test hits the
DCHECK without the asynchronous LoadComplete.

Bug: 1275600
Change-Id: I17ee35c26af95a3ba79a79da75107087c767eef2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3489148
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#977783}