Update existing media history scores with an expiration time
This CL updates existing media history scores with an expiration time.
The CL increments the schema version (even though technically there are
no changes to the schema) to indicate whether the existing media history
scores should be re-committed with an expiration time.

This CL also increases the expiration of media history scores from 30
days to 90 days to be consistent with when history is automatically
expired.

Bug: 1267325
Change-Id: I2f85719719d2c5e2705c7a886433cd86f8026b77
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3501173
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Evan Liu <evliu@google.com>
Cr-Commit-Position: refs/heads/main@{#977838}