[Journeys] DeLorean on Android JNI bridge and feature flag
This adds a JNI bridge and feature flag for plumbing history clusters
data through to Java which allows for local testing. Followup CLs
will add the UI to which this data is bound.

Bug: 1303171

Change-Id: I07f77fe37880ab167661303c38a48cadd964e9a6

significant chunk of previously stripped feature code for generating
history clusters

Binary-Size: Size increase is the unavoidable result of including a
Change-Id: I07f77fe37880ab167661303c38a48cadd964e9a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3450997
Reviewed-by: Tommy Li <tommycli@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Commit-Queue: Patrick Noland <pnoland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977853}