[SH 2.0 iOS] Cache last text fragment for resharing
If the user clicks on a same-document link, this removes the text
fragment from the URL. Currently, we simply reshare the last committed
URL, so a user attempting to share a highlight after a same-doc
navigation will instead share a link with no highlights.

This patch caches the text fragments that caused a successful highlight,
and sends them to the browser layer when the user taps a highlight. If
the user then selects reshare, the fragments received from JS are
appended to the current URL, and the result is shared.

These need to be cached in JS, rather than in the browser layer, because
"stale" highlights may still be present on a page reached by back-
navigation. This approach ties the lifecycle of the cached fragment data
to the lifecycle of the marks actually rendered on the page.

Bug: 1298225
Change-Id: I4e2817f0f71fa733beae0d94f41753b1f2fc4106
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3501933
Reviewed-by: Guillaume Jenkins <gujen@google.com>
Commit-Queue: Tommy Martino <tmartino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977875}