blink/transitions: Ensure correct root snapshot with scrollbar.
When the root element has dimensions which introduce a scrollbar, the
paint effect node associated with it generates multiple cc effect nodes.
Each of these effect nodes generates an independent render surface which
then end up sharing the same resource id for shared elements (which is
1:1 with the paint effect node).

The fundamental issue here is that for the root snapshot we assume
that all paint chunks for the LayoutView, except the ::transition
element, can be grouped together under a single paint effect node. This
is done by reparenting the ::transition element above LayoutView's effect
node.

But the assumption above is incorrect since the scrollbar (painted as
a part of LayoutView) also needs to be excluded from this snapshot. This
change assumes that the first cc effect node generated for the
LayoutView's paint effect node groups all content that should be captured.

The long term fix needs to clearly define which content and effects for
the LayoutView should be included in the snapshot and introduce a
grouping paint effect node to be used only for that content.

R=pdr@chromium.org, vmpstr@chromium.org

Bug: 1303081
Change-Id: I47faf618afa1c951b7e4648f77fc8ee6c7cbfb3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3499792
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Auto-Submit: Khushal Sagar <khushalsagar@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Khushal Sagar <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977900}