Refactor bubble border in TrayBubbleView
This class creates a pointer `bubble_border_` and an unique_ptr that
point to the same thing, then later on transfer ownership of the
unique_ptr and keep using the pointer `bubble_border_`. Ideally, we
should avoid this. After move and when we need to use it, we should
retrieve the pointer from the new owners.

Change-Id: I6672b0c31382da4e0da077800a2703ea9e60150e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3498973
Reviewed-by: Dana Fried <dfried@chromium.org>
Reviewed-by: Ahmed Mehfooz <amehfooz@chromium.org>
Commit-Queue: Andre Le <leandre@chromium.org>
Cr-Commit-Position: refs/heads/main@{#977913}