[CrOS Bluetooth] Clear state changes when adapter becomes unavailable.
Update AdapterStateController to clear any in-progress and queued state
changes when the adapter becomes not present. Prior to this, if the
adapter became unavailable during a state changed and the SetPowered()
callback is never resolved, any new state change requests will be
blocked by the never-resolved in-progress state change, even if the
adapter becomes present again.

Bug: b/219596007
Change-Id: Icf5f512ce86b92bf279afbc8ecdfe5988d5e58e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3504103
Reviewed-by: Chad Duffin <chadduffin@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Gordon Seto <gordonseto@google.com>
Cr-Commit-Position: refs/heads/main@{#977929}