arc: Close DocumentsProvider FDs after FileStreamWriter operations
Currently openFileToWrite in ARC detaches the Android file descriptor
and causes DocumentsProvider app's onClose() to be called immediately
before the actual copy / write file operation(s) even started. This in
turn results in a FileDescriptorDetachedException.

This issue was reported by JottaCloud and triaged by Chrome OS DevRel.
The fix is to sessionize the open operation on Android side so that
the originally opened ParcelFileDescriptor can be closed after the
Chromium side file write operation(s). For FileStreamWriter, all file
operations are completed when the writer object is destroyed.

The Android CLs counterpart are ag/16963727 and ag/16963726 (P and R).

BUG=b:215364617
TEST=./out/Debug/unit_tests --gtest_filter=ArcFileSystemOperationRunnerTest*/ArcContentFileSystemFileStreamWriterTest*
TEST=./out/Debug/browser_tests --gtest_filter=QuickView/FilesAppBrowserTest.Test/openQuickViewDocumentsProvider*
TEST=Manually verified using bug's StorageProviderWriteException APK

Change-Id: I91a5b5e064ef4fc8278ef4d9b69bf0e856c9712f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3476721
Reviewed-by: David Jacobo <djacobo@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Alan Ding <alanding@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978052}