Add a new feature for BFCache time to live.
This allows us to experiment with the TTL for cached pages separately
from the main BFCache feature - without this feature, changing the TTL
requires changing the main BFCache feature experiment.

This CL also adds helper methods to the BFCache class that return what
are effectively 'standard' default parameters to use when testing with
the BFCache, and shows their usage in document_service_browsertest and
devtools_issue_storage_browsertest.

As a note, this new feature is in addition to an existing TTL parameter on
the main BackForwardCache feature. That parameter is used in ~30
browsertests so replacing it in one CL is impractical. In this CL we
support both parameters, with the newer parameter taking precedence.

Change-Id: I2966f2fcd35a80dc429cde32924675384f2df5c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3435292
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Fergal Daly <fergal@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Commit-Queue: Darren Willis <djw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978134}