Add timeout to start ambient UI with at least minimum topics available.
This should have no impact on the current slideshow screensaver, but
for the animated screensaver, we can start the UI with just 1 topic in
the model. In the worst case scenario where the network is very
congested, we should start the UI after a timeout (10 seconds) if
we have the minimum available.

Bug: b:222239974
Change-Id: I6b0f2c8ea65f19547fd5b4a3895f8472a02e0d5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3499031
Reviewed-by: Tao Wu <wutao@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Commit-Queue: Eric Sum <esum@google.com>
Cr-Commit-Position: refs/heads/main@{#978202}