Change threshold for chunking user policies
Change _LAST_TOP_LEVEL_POLICY_ID from 979 to 1015, i.e. start chunking
user policies once we reach ID number 1016. The reason for this change
is to delay the point in time when we start chunking in order to allow
for more time for backend changes to be implemented.

Bug: 1237044, b:217516275
Test: python3 generate_policy_source_test.py
Change-Id: Iff17167991a6ec83e75b693c45584af2e707f39c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3503910
Reviewed-by: Alexander Hendrich <hendrich@chromium.org>
Commit-Queue: Jana Grill <janagrill@google.com>
Cr-Commit-Position: refs/heads/main@{#978205}