[Lacros] Implement Lacros-compatible GetDeviceType
Currently, Lacros is not aware of the type of device ("Chromebook",
"Chromebox" etc.) it is running on. However, this information is needed
for some UI elements.

This CL uses the Lacros init_params to tell Lacros which type of device
it is running on. It implements a GetDeviceType() which can be used by
both Ash and Lacros - this is based on the existing GetDeviceType() in
ash/constants/devicetype.cc.
Current uses of the ash-exclusive GetDeviceType() will be migrated to
use the new one in follow-up CLs.

Bug: 1284585
Change-Id: Iceb0b2e10744657e18a2555383a0fc57d81e65f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3477020
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Fabian Sommer <fabiansommer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978212}