chromebox_for_meetings: Prevent App Bootloop
Always map the pending_remote for
chromeos::cfm::mojom::CfmServiceContext and reset it with a known code
and description for none-cfm devices and when the feature flag is
disabled. This is a remediation for the boot loop issue experienced by
trusted kiosked applications when attempting to bind to the
CfmServiceContext due to intended behaviour by chrome to terminate
processes who request to bind to mojom interfaces that are not found in
the interface_binders mapping.


./out/Default/chromeos_unittests --dbus-stub --single-process-tests
--gtest_filter=Cfm*Test.*"

Bug: 1301371
Test: "autoninja -C out/Default chromeos:chromeos_unittests &&
Change-Id: I28dbbae6df599883af0b68f13dc3ae065d54a1db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3495465
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Kyle Williams <kdgwill@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978217}