[iOS] Fix logic error in page change logging in tab grid.
The tab grid resets the pageChangeInteraction value after it logs a
page change. If the user is still dragging the scroll view and triggers
a second page change, this means that pageChangeInteraction will be None
when the second page change is logged. This is a NOTREACHED condition
that's still reachable (hence the bug).

This CL removes the pageChangeInteraction reset at the end of the
logging method. The pageChangeInteraction is reset after scrolling or
dragging/tapping the page control is complete.

tabs and then (in the same gesture) back to the regular tabs. There
should be no crash.

TEST: In canary, enter the tab grid and slowly scroll to the incongnito
Fixed: 1003724
Change-Id: If755aef178b9b4f3e3786fc23646bc4ead337696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3503633
Reviewed-by: David Jean <djean@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978220}