[iOS] Fix doubled events for page control taps.
The tab grid page control is sending UIControlTouchUpInside events
twice each time it's tapped. This is because taps are handled by a
gesture recognizer that has cancelsTouched = YES (the default). Since
the control should still handle cancellations of non-tap touches
correctly, there still needs to be an impmlementation of
-touchesCancelled:.

This CL hangs on to a pointer to the tap recognizer, and then checks
its state in -touchesCancelled:. Control events are not sent if the
tap recognizer is in its 'ending' state (which is where it cancels
touches).

Fixed: 935403
Change-Id: Ief75be7059735d9718be2b532947e5c9f6578b7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3498107
Reviewed-by: David Jean <djean@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978273}