Port omnibox chip to color pipeline.
This is a slight behavioral change.  For the "normal visibility" chip,
the colors used are now the button foreground/background colors instead
of some one-off hardcoded colors.  There are two primary effects of
this:
* In dark mode, the chip background gets slightly darker.
* When the Windows "pervasive system accent color" experiment is
  enabled, the chip text follows the accent color change instead of
  remaining blue.

This ports directly to color pipeline instead of sticking with the
ThemeService system, since achieving the above with the old theme system
is slightly tricky (there is no theme color corresponding to "the accent
color", for example).

Bug: 1296206
Change-Id: Iebb0ba66e5041deea56410edd1b9350ff183bc58
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3504184
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Olesia Marukhno <olesiamarukhno@google.com>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978276}