Android: Specify device when running benchmark.py
When running benchmarks for emulators, specify the device.serial to
apk_operations.py so that the physical devices (if any) do not need to
be unplugged and are not used during install. This avoids the error in
apk_operations.py which requires either `--all` to be passed or
`--device` to be passed when multiple devices are connected.

Bug: 663492
Change-Id: Ib1f1a6d1ce5dff21ca8bdf8cb9e8affc3ab51a28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3507194
Commit-Queue: Peter Wen <wnwen@chromium.org>
Auto-Submit: Peter Wen <wnwen@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978281}