[iOS] Improve webpage context menu JavaScript
This is the first cl in a series. It refactors the code that loops
through elements to find targets at locations on the page to consider
not just the top element and its ancestors but also other
elements behind it. Future cls will add support for image selection
for div backgrounds, transparent overlays, and opaque layers blocking
image selection.

Bug: 1295930
Change-Id: I5a2a1284d42528a3a4d1e77a6b11102d9fefc712
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3500911
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Jason Hu <hujasonx@google.com>
Cr-Commit-Position: refs/heads/main@{#978351}