Reuse SQL fragments and column-extraction code for attribution storage
In particular, we use macros to share column lists, and we enable
ReadReportFromStatement to call ReadSourceFromStatement, instead of
extracting source columns itself.

This makes the queries less prone to error and avoids nearly duplicated
but slightly different code in multiple places in this file.

Change-Id: I71ba9670bf21ac801629bdc5ab9aff58dadc3fbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3501935
Reviewed-by: Nan Lin <linnan@chromium.org>
Reviewed-by: John Delaney <johnidel@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978363}