Toggle annotation layer off when the toolbar is being closed
The recording can end when the annotation layer is still enabled, so we
should manually turn it off if it is still enabled. Also update the
annotation tray icon because it will carry over the old state into the
next annotation session otherwise.

Tested:Made sure the full screen recordings are WAI, and that Chrome
does not crash when ending the recording for window/region recordings.

Bug: b/222343206
Change-Id: I065c568b0ef0849cbaf3aab7638314bbc50ba313
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3503098
Reviewed-by: Li Lin <llin@chromium.org>
Commit-Queue: Courtney Wong <courtneywong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978447}