Migrate stale tabs to CriticalPersistedTabData
It has been observed that stale Tabs (those which sit uninteracted
with) are not being migrated to CriticalPersistedTabData. This is
because we only migrate a Tab to CriticalPersistedTabData if a Tab
changes (e.g. URL changes). This CL introduces a mechanism to migrate
5 Tabs to CriticalPersistedTabData for every 1 save of TabState for a
Tab. This approach was taken with the following considerations:
- Tabs should not be migrated on startup as this will risk a startup
  regression. After a TabState file is saved is a time of relatively
  low resource utilization.
- Migrating too many Tabs at a time should be avoided such that the
  system doesn't get flooded with saves (some users have 100+ Tabs).
  5 seems reasonable for a Batch size.

Bug: 1303027
Change-Id: If1d0159e9f4f0e799e08a771e7b9b7064ce3f7cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3498958
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: David Maunder <davidjm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#978453}